Add current development project information and update portfolio page
Added a section to showcase current development projects on the front page. Also made several modifications to the portfolio page to enhance UX. Changes include adding an indication of the current date, refining the look and feel of timeline components and improving the style of project cards. Updated several module versions for better performance and stability.